chore: [SEISMO-0138] Setup for development environment

+++ routes/notifications.js
@@ -1,46 +1,51 @@
 const express = require('express');
 const Subscription = require('../models/subscription');
-const mongodb = require('../services/mongodb')
 const events = require('../services/events')
+const mongoose = require('mongoose');
 const redis = require("redis")
 const webpush = require('web-push')
-const config = require('../config')
+require('dotenv').config({path: __dirname + '/../.env'})
 
-mongodb.connect();
 
 const router = express.Router();
 router.post('/subscribe', async (req,res) => {
-  const subscriptionRequest = req.body // TODO: Validate request body
-  const subExists = await Subscription.exists(
-    {endpoint: subscriptionRequest.endpoint}
-  )
-  if(subExists){
-    console.log('Old subscription found')
+  if(mongoose.connection.readyState === 1) { // connected to MongoDB
+    const subscriptionRequest = req.body // TODO: Validate request body
+    const subExists = await Subscription.exists(
+      {endpoint: subscriptionRequest.endpoint}
+    )
+    if(subExists){
+      console.log('Old subscription found')
+    }else{
+      console.log(subExists)
+      const subscription = new Subscription(subscriptionRequest);
+        const savedSubscription = await subscription.save();
+        console.log('New subscription created')
+        res.status(201).json({'success': true}) // send 201 - resource created
+    }
   }else{
-    console.log(subExists)
-    console.log('New subscription created')
-    const subscription = new Subscription(subscriptionRequest); 
-    const savedSubscription = await subscription.save();
+    console.warn("Can't create new subscription, MongoDB not connected");
+    res.status(500).json({'success': false}) // send 500 - resource not created
   }
-
-  // send 201 - resource created
-  res.status(201).json({'success': true})
 })
 
 
 // redis real-time comms with SC, and broadcasts notifs
-const proxy = () =>{
-  const redisChannel = "SC_*" // SC_PICK or SC_EVENT
-  const redisClient = redis.createClient(config.redis)
-  redisClient.psubscribe(redisChannel)
-
+var subscriber;
+const redisProxy = async (new_config) =>{
   webpush.setVapidDetails(
     process.env.WEB_PUSH_CONTACT,
     process.env.PUBLIC_VAPID_KEY,
     process.env.PRIVATE_VAPID_KEY,
-  )
-  redisClient.on("pmessage", async (pattern, channel, message) => {
-    if(channel === "SC_EVENT"){
+  );
+
+  try {
+    subscriber = redis.createClient(new_config ? new_config : {
+      url:`redis://${process.env.REDIS_HOST}:${process.env.REDIS_PORT}`
+    });
+    await subscriber.connect()
+    await subscriber.pSubscribe("SC_EVENT", async (message, channel) => {
+      console.log(`notifications.js received: ${channel}`)
       message = JSON.parse(message)
 
       if(message.magnitude_value > 5.5){
@@ -57,26 +62,36 @@ const proxy = () =>{
           body: `Magnitude ${updatedEvent.magnitude_value} in ${address}`,
         })
 
-        const subscribers = await Subscription.find({});
-        subscribers.forEach(subscriber => {
-          webpush.sendNotification(subscriber, payload)
-            .then(console.log(`Sent notif to ${subscriber._id}`))
-            .catch(response => {
-              switch(response.statusCode){
-                case 400:
-                case 410:
-                  console.log(`Subscription gone for ${subscriber._id}`)
-                  Subscription.deleteOne(subscriber)
-                  .then(console.log(`Deleted ${subscriber.id}`))
-                  break;
-                default:
-                  console.log(JSON.parse(e.body))
-              }
-            })
-        })
+        if(mongoose.connection.readyState === 1) { // connected to MongoDB
+          const subscribers = await Subscription.find({});
+          subscribers.forEach(subscriber => {
+            webpush.sendNotification(subscriber, payload)
+              .then(console.log(`Sent notif to ${subscriber._id}`))
+              .catch(response => {
+                switch(response.statusCode){
+                  case 400:
+                  case 410:
+                    console.log(`Subscription gone for ${subscriber._id}`)
+                    Subscription.deleteOne(subscriber)
+                    .then(console.log(`Deleted ${subscriber.id}`))
+                    break;
+                  default:
+                    console.log(JSON.parse(e.body))
+                }
+              })
+          })
+        }else{
+          console.warn("Can't access subscriptions, MongoDB not connected");
+        }
       }
-    }
-  });
+    });
+  } catch (err) {
+    console.trace(`In Redis setup...\n ${err}`)
+    //TODO: properly handle this scenario
+  }
+}
+const quitRedisProxy = async () => {
+  subscriber && (await subscriber.quit())
 }
 
-module.exports = {router, proxy}
+module.exports = {router, redisProxy, quitRedisProxy}
